requirements.txt
Flask
XlsxWriter
Flask-WTF
wtforms-html5
inflect
flask-login
gunicorn
pyPDF2
flask-sqlalchemy
reportlab
